Continued work on making tabview a widget. Unfinished.
